refactor(sdk): group per-contract bindings under packages/
